# openapi/auth.yaml
openapi: 3.0.3
info:
  title: Yarah Authentication API
  version: 2.0.0
  description: Authentication endpoints with separated auth and profile tables
paths:
...
  /api/auth/users:
    post:
      summary: Register new user
      description: |
        Creates a new user account. When email verification is required, a delivery failure is returned as an error after the account is created. Retry delivery with POST /api/auth/email/send-verification instead of registering the same email again.
      tags:
          - Client
      parameters:
        - name: client_type
          in: query
          schema:
            type: string
            enum: [web, mobile, desktop, server]
            default: web
          description: |
            Client type determines how refresh tokens are returned:
            - web: Refresh token stored in httpOnly cookie, csrfToken returned in response
            - mobile/desktop/server: refreshToken returned directly in response body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  description: Password meeting configured requirements (check /api/auth/email/config for current requirements)
                  example: securepassword123
                name:
                  type: string
                  example: John Doe
                redirectTo:
                  type: string
                  format: uri
                  description: >-
                    Used for link-based email verification. The email link always
                    opens an Yarah backend endpoint first; after the token is
                    verified, Yarah redirects the browser to this URL. This URL
                    must be included in allowedRedirectUrls. Recommended: use your
                    app's sign-in page.
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  accessToken:
                    type: string
                    nullable: true
                    description: JWT authentication token (null if email verification required)
                  csrfToken:
                    type: string
                    nullable: true
                    description: CSRF token for use with refresh endpoint (web clients only, null if email verification required)
                  refreshToken:
                    type: string
                    nullable: true
                    description: Refresh token for mobile/desktop/server clients (null for web clients or if email verification required)
                  requireEmailVerification:
                    type: boolean
                    description: Whether email verification is required before login
        '400':
          description: Invalid request
        '403':
          description: Signups disabled — project has disabled new user registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: AUTH_SIGNUP_DISABLED
                message: User signups are disabled for this project.
                statusCode: 403
        '409':
          description: User already exists
        '429':
          description: User created, but verification email delivery was rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: AUTH_VERIFICATION_EMAIL_DELIVERY_FAILED
                message: The user account was created, but the verification email could not be sent.
                statusCode: 429
                nextActions: The user account already exists. Retry delivery with POST /api/auth/email/send-verification instead of registering again.
        '500':
          description: User created, but the verification email could not be sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: AUTH_VERIFICATION_EMAIL_DELIVERY_FAILED
                message: The user account was created, but the verification email could not be sent.
                statusCode: 500
                nextActions: The user account already exists. Retry delivery with POST /api/auth/email/send-verification instead of registering again.

    get:
      summary: List all users (admin only)
      description: Returns paginated list of users
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: string
            default: '0'
          description: Number of records to skip
        - name: limit
          in: query
          schema:
            type: string
            default: '10'
          description: Maximum number of records to return
        - name: search
          in: query
          schema:
            type: string
          description: Search by email or name
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  pagination:
                    type: object
                    properties:
                      offset:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    delete:
      summary: Delete users (admin only)
      description: Delete multiple users by their IDs
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
              required:
                - userIds
      responses:
        '200':
          description: Users deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedCount:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
...
  /api/auth/oauth/configs:
    get:
      summary: List all OAuth configurations
      description: Get all configured OAuth providers (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of OAuth configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OAuthConfig'
                  count:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    post:
      summary: Create OAuth configuration
      description: Create a new OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
              properties:
                provider:
                  type: string
                  enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
                clientId:
                  type: string
                nativeClientIds:
                  type: array
                  maxItems: 20
                  description: Server-trusted native app client IDs used as ID-token audiences
                  items:
                    type: string
                    minLength: 1
                    maxLength: 255
                clientSecret:
                  type: string
                redirectUri:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                useSharedKey:
                  type: boolean
      responses:
        '200':
          description: OAuth configuration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthConfig'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
...
  /api/auth/oauth/{provider}:
    get:
      summary: Initiate OAuth flow (PKCE)
      description: |
        Generate OAuth authorization URL for any supported provider using PKCE flow.

        For mobile/desktop/server clients using PKCE:
        1. Generate code_verifier (random string, 43-128 chars)
        2. Generate code_challenge = Base64URL(SHA256(code_verifier))
        3. Call this endpoint with code_challenge
        4. After OAuth callback, use /api/auth/oauth/exchange with code_verifier
      tags:
        - Client
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication (receives yarah_code parameter)
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
          description: |
            PKCE code challenge for mobile/desktop/server clients.
            Generate using: Base64URL(SHA256(code_verifier))
            Required for secure token exchange in native apps or trusted server-side clients.
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                    format: uri
                    description: URL to redirect user for OAuth provider login
        '400':
          description: Invalid request or provider not supported
        '500':
          description: OAuth not configured
...
  /api/auth/oauth/shared/callback/{state}:
    get:
      summary: Shared OAuth callback handler
      description: Handles OAuth callbacks from Yarah Cloud shared OAuth
      tags:
        - Client
      parameters:
        - name: state
          in: path
          required: true
          schema:
            type: string
          description: JWT state parameter
        - name: success
          in: query
          schema:
            type: string
          description: Success flag
        - name: error
          in: query
          schema:
            type: string
          description: Error message
        - name: payload
          in: query
          schema:
            type: string
          description: Base64 encoded user payload
      responses:
        '302':
          description: Redirect to application with access token or error
          headers:
            Location:
              schema:
                type: string
                format: uri
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        profile:
          type: object
          nullable: true
          additionalProperties: true
          description: User profile data (name, avatar_url, and custom fields)
          properties:
            name:
              type: string
            avatar_url:
              type: string
              format: uri
        metadata:
          type: object
          nullable: true
          additionalProperties: true
          description: System metadata (device ID, login IP, etc.)
        emailVerified:
          type: boolean
        providers:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OAuthConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
        clientId:
          type: string
          nullable: true
        nativeClientIds:
          type: array
          maxItems: 20
          description: Server-trusted native app client IDs used as ID-token audiences
          items:
            type: string
            minLength: 1
            maxLength: 255
        redirectUri:
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
          nullable: true
        useSharedKey:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - statusCode
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Email is already in use"
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        nextActions:
          type: string
          description: Suggested action to resolve the error
          example: "Please use a different email address"
...